// src/api/generated/types.ts
export type AL_AnimeDetailsById_Media = {
    averageScore?: number
    characters?: AL_AnimeDetailsById_Media_Characters
    description?: string
    duration?: number
    endDate?: AL_AnimeDetailsById_Media_EndDate
    genres?: Array<string>
    id: number
    meanScore?: number
    popularity?: number
    rankings?: Array<AL_AnimeDetailsById_Media_Rankings>
    recommendations?: AL_AnimeDetailsById_Media_Recommendations
    relations?: AL_AnimeDetailsById_Media_Relations
    siteUrl?: string
    staff?: AL_AnimeDetailsById_Media_Staff
    startDate?: AL_AnimeDetailsById_Media_StartDate
    studios?: AL_AnimeDetailsById_Media_Studios
    trailer?: AL_AnimeDetailsById_Media_Trailer
}

/**
 * - Filepath: internal/api/anilist/client_gen.go
 * - Filename: client_gen.go
 * - Package: anilist
 */
export type AL_AnimeDetailsById_Media_Characters = {
    edges?: Array<AL_AnimeDetailsById_Media_Characters_Edges>
}

/**
 * - Filepath: internal/api/anilist/client_gen.go
 * - Filename: client_gen.go
 * - Package: anilist
 */
export type AL_AnimeDetailsById_Media_Characters_Edges = {
    id?: number
    name?: string
    node?: AL_BaseCharacter
    role?: AL_CharacterRole
}

/**
 * - Filepath: internal/api/anilist/client_gen.go
 * - Filename: client_gen.go
 * - Package: anilist
 */
export type AL_AnimeDetailsById_Media_EndDate = {
    day?: number
    month?: number
    year?: number
}

/**
 * - Filepath: internal/api/anilist/client_gen.go
 * - Filename: client_gen.go
 * - Package: anilist
 */
export type AL_AnimeDetailsById_Media_Rankings = {
    allTime?: boolean
    context: string
    format: AL_MediaFormat
    rank: number
    season?: AL_MediaSeason
    type: AL_MediaRankType
    year?: number
}

/**
 * - Filepath: internal/api/anilist/client_gen.go
 * - Filename: client_gen.go
 * - Package: anilist
 */
export type AL_AnimeDetailsById_Media_Recommendations = {
    edges?: Array<AL_AnimeDetailsById_Media_Recommendations_Edges>
}

/**
 * - Filepath: internal/api/anilist/client_gen.go
 * - Filename: client_gen.go
 * - Package: anilist
 */
export type AL_AnimeDetailsById_Media_Recommendations_Edges = {
    node?: AL_AnimeDetailsById_Media_Recommendations_Edges_Node
}
// ...
export type AL_AnimeDetailsById_Media_Relations = {
    edges?: Array<AL_AnimeDetailsById_Media_Relations_Edges>
}

/**
 * - Filepath: internal/api/anilist/client_gen.go
 * - Filename: client_gen.go
 * - Package: anilist
 */
export type AL_AnimeDetailsById_Media_Relations_Edges = {
    node?: AL_BaseAnime
    relationType?: AL_MediaRelation
}

/**
 * - Filepath: internal/api/anilist/client_gen.go
 * - Filename: client_gen.go
 * - Package: anilist
 */
export type AL_AnimeDetailsById_Media_Staff = {
    edges?: Array<AL_AnimeDetailsById_Media_Staff_Edges>
}

/**
 * - Filepath: internal/api/anilist/client_gen.go
 * - Filename: client_gen.go
 * - Package: anilist
 */
export type AL_AnimeDetailsById_Media_Staff_Edges = {
    node?: AL_AnimeDetailsById_Media_Staff_Edges_Node
    role?: string
}
// ...
export type AL_AnimeDetailsById_Media_StartDate = {
    day?: number
    month?: number
    year?: number
}

/**
 * - Filepath: internal/api/anilist/client_gen.go
 * - Filename: client_gen.go
 * - Package: anilist
 */
export type AL_AnimeDetailsById_Media_Studios = {
    nodes?: Array<AL_AnimeDetailsById_Media_Studios_Nodes>
}

/**
 * - Filepath: internal/api/anilist/client_gen.go
 * - Filename: client_gen.go
 * - Package: anilist
 */
export type AL_AnimeDetailsById_Media_Studios_Nodes = {
    id: number
    name: string
}

/**
 * - Filepath: internal/api/anilist/client_gen.go
 * - Filename: client_gen.go
 * - Package: anilist
 */
export type AL_AnimeDetailsById_Media_Trailer = {
    id?: string
    site?: string
    thumbnail?: string
}
// ...
export type AL_BaseAnime = {
    id: number
    idMal?: number
    siteUrl?: string
    status?: AL_MediaStatus
    season?: AL_MediaSeason
    type?: AL_MediaType
    format?: AL_MediaFormat
    seasonYear?: number
    bannerImage?: string
    episodes?: number
    synonyms?: Array<string>
    isAdult?: boolean
    countryOfOrigin?: string
    meanScore?: number
    description?: string
    genres?: Array<string>
    duration?: number
    trailer?: AL_BaseAnime_Trailer
    title?: AL_BaseAnime_Title
    coverImage?: AL_BaseAnime_CoverImage
    startDate?: AL_BaseAnime_StartDate
    endDate?: AL_BaseAnime_EndDate
    nextAiringEpisode?: AL_BaseAnime_NextAiringEpisode
}

/**
 * - Filepath: internal/api/anilist/client_gen.go
 * - Filename: client_gen.go
 * - Package: anilist
 */
export type AL_BaseAnime_CoverImage = {
    color?: string
    extraLarge?: string
    large?: string
    medium?: string
}

/**
 * - Filepath: internal/api/anilist/client_gen.go
 * - Filename: client_gen.go
 * - Package: anilist
 */
export type AL_BaseAnime_EndDate = {
    day?: number
    month?: number
    year?: number
}

/**
 * - Filepath: internal/api/anilist/client_gen.go
 * - Filename: client_gen.go
 * - Package: anilist
 */
export type AL_BaseAnime_NextAiringEpisode = {
    airingAt: number
    episode: number
    timeUntilAiring: number
}

/**
 * - Filepath: internal/api/anilist/client_gen.go
 * - Filename: client_gen.go
 * - Package: anilist
 */
export type AL_BaseAnime_StartDate = {
    day?: number
    month?: number
    year?: number
}

/**
 * - Filepath: internal/api/anilist/client_gen.go
 * - Filename: client_gen.go
 * - Package: anilist
 */
export type AL_BaseAnime_Title = {
    english?: string
    native?: string
    romaji?: string
    userPreferred?: string
}

/**
 * - Filepath: internal/api/anilist/client_gen.go
 * - Filename: client_gen.go
 * - Package: anilist
 */
export type AL_BaseAnime_Trailer = {
    id?: string
    site?: string
    thumbnail?: string
}
// ...
export type AL_MediaFormat = "TV" |
    "TV_SHORT" |
    "MOVIE" |
    "SPECIAL" |
    "OVA" |
    "ONA" |
    "MUSIC" |
    "MANGA" |
    "NOVEL" |
    "ONE_SHOT"

/**
 * - Filepath: internal/api/anilist/models_gen.go
 * - Filename: models_gen.go
 * - Package: anilist
 * @description
 *  Media list watching/reading status enum.
 */
export type AL_MediaListStatus = "CURRENT" |
    "PLANNING" |
    "COMPLETED" |
    "DROPPED" |
    "PAUSED" |
    "REPEATING"

/**
 * - Filepath: internal/api/anilist/models_gen.go
 * - Filename: models_gen.go
 * - Package: anilist
 * @description
 *  The type of ranking
 */
export type AL_MediaRankType = "RATED" | "POPULAR"
// ...
export type AL_MediaSeason = "WINTER" | "SPRING" | "SUMMER" | "FALL"
// ...
export type AL_MediaStatus = "FINISHED" | "RELEASING" | "NOT_YET_RELEASED" | "CANCELLED" | "HIATUS"

/**
 * - Filepath: internal/api/anilist/models_gen.go
 * - Filename: models_gen.go
 * - Package: anilist
 * @description
 *  Media type enum, anime or manga.
 */
export type AL_MediaType = "ANIME" | "MANGA"
// ...
export type Anime_Entry = {
    mediaId: number
    media?: AL_BaseAnime
    listData?: Anime_EntryListData
    libraryData?: Anime_EntryLibraryData
    downloadInfo?: Anime_EntryDownloadInfo
    episodes?: Array<Anime_Episode>
    nextEpisode?: Anime_Episode
    localFiles?: Array<Anime_LocalFile>
    anidbId: number
    currentEpisodeCount: number
    _isNakamaEntry: boolean
    nakamaLibraryData?: Anime_NakamaEntryLibraryData
}

/**
 * - Filepath: internal/library/anime/entry_download_info.go
 * - Filename: entry_download_info.go
 * - Package: anime
 */
export type Anime_EntryDownloadEpisode = {
    episodeNumber: number
    aniDBEpisode: string
    episode?: Anime_Episode
}

/**
 * - Filepath: internal/library/anime/entry_download_info.go
 * - Filename: entry_download_info.go
 * - Package: anime
 */
export type Anime_EntryDownloadInfo = {
    episodesToDownload?: Array<Anime_EntryDownloadEpisode>
    canBatch: boolean
    batchAll: boolean
    hasInaccurateSchedule: boolean
    rewatch: boolean
    absoluteOffset: number
}

/**
 * - Filepath: internal/library/anime/entry_library_data.go
 * - Filename: entry_library_data.go
 * - Package: anime
 */
export type Anime_EntryLibraryData = {
    allFilesLocked: boolean
    sharedPath: string
    unwatchedCount: number
    mainFileCount: number
}

/**
 * - Filepath: internal/library/anime/entry.go
 * - Filename: entry.go
 * - Package: anime
 */
export type Anime_EntryListData = {
    progress?: number
    score?: number
    status?: AL_MediaListStatus
    repeat?: number
    startedAt?: string
    completedAt?: string
}

/**
 * - Filepath: internal/library/anime/episode.go
 * - Filename: episode.go
 * - Package: anime
 */
export type Anime_Episode = {
    type: Anime_LocalFileType
    /**
     * e.g, Show: "Episode 1", Movie: "Violet Evergarden The Movie"
     */
    displayTitle: string
    /**
     * e.g, "Shibuya Incident - Gate, Open"
     */
    episodeTitle: string
    episodeNumber: number
    /**
     * AniDB episode number
     */
    aniDBEpisode?: string
    absoluteEpisodeNumber: number
    /**
     * Usually the same as EpisodeNumber, unless there is a discrepancy between AniList and AniDB
     */
    progressNumber: number
    localFile?: Anime_LocalFile
    /**
     * Is in the local files
     */
    isDownloaded: boolean
    /**
     * (image, airDate, length, summary, overview)
     */
    episodeMetadata?: Anime_EpisodeMetadata
    /**
     * (episode, aniDBEpisode, type...)
     */
    fileMetadata?: Anime_LocalFileMetadata
    /**
     * No AniDB data
     */
    isInvalid: boolean
    /**
     * Alerts the user that there is a discrepancy between AniList and AniDB
     */
    metadataIssue?: string
    baseAnime?: AL_BaseAnime
    _isNakamaEpisode: boolean
}

/**
 * - Filepath: internal/library/anime/episode_collection.go
 * - Filename: episode_collection.go
 * - Package: anime
 */
export type Anime_EpisodeCollection = {
    hasMappingError: boolean
    episodes?: Array<Anime_Episode>
    metadata?: Metadata_AnimeMetadata
}

/**
 * - Filepath: internal/library/anime/episode.go
 * - Filename: episode.go
 * - Package: anime
 */
export type Anime_EpisodeMetadata = {
    anidbId?: number
    image?: string
    airDate?: string
    length?: number
    summary?: string
    overview?: string
    isFiller?: boolean
    /**
     * Indicates if the episode has a real image
     */
    hasImage?: boolean
    title?: string
}
// ...
export type Anime_LocalFile = {
    path: string
    name: string
    parsedInfo?: Anime_LocalFileParsedData
    parsedFolderInfo?: Array<Anime_LocalFileParsedData>
    metadata?: Anime_LocalFileMetadata
    locked: boolean
    /**
     * Unused for now
     */
    ignored: boolean
    mediaId: number
}

/**
 * - Filepath: internal/library/anime/localfile.go
 * - Filename: localfile.go
 * - Package: anime
 */
export type Anime_LocalFileMetadata = {
    episode: number
    aniDBEpisode: string
    type: Anime_LocalFileType
}

/**
 * - Filepath: internal/library/anime/localfile.go
 * - Filename: localfile.go
 * - Package: anime
 */
export type Anime_LocalFileParsedData = {
    original: string
    title?: string
    releaseGroup?: string
    season?: string
    seasonRange?: Array<string>
    part?: string
    partRange?: Array<string>
    episode?: string
    episodeRange?: Array<string>
    episodeTitle?: string
    year?: string
}

/**
 * - Filepath: internal/library/anime/localfile.go
 * - Filename: localfile.go
 * - Package: anime
 */
export type Anime_LocalFileType = "main" | "special" | "nc"

/**
 * - Filepath: internal/library/anime/missing_episodes.go
 * - Filename: missing_episodes.go
 * - Package: anime
 */
export type Anime_MissingEpisodes = {
    episodes?: Array<Anime_Episode>
    silencedEpisodes?: Array<Anime_Episode>
}

/**
 * - Filepath: internal/library/anime/entry_library_data.go
 * - Filename: entry_library_data.go
 * - Package: anime
 */
export type Anime_NakamaEntryLibraryData = {
    unwatchedCount: number
    mainFileCount: number
}